Save the params from the launch and use some as metadata on create
Added the model AppLaunch to store the params received in the launch.
The user object now stores the launch nonce, that is used to index the
launches in the database. When creating a conference, the launch data
is loaded from the database and some of it is set as metadata.