New page where users go when waiting for a conference to start
Instead of just showing a message in rooms/show, when trying to join the
user will be redirected to a wait page that opens in a new tab.
This page has a websocket to receive an update and automatically join the
user when the conference starts.
This is the only page using ActionCable now, removed the default channel
that was being started in all pages unnecessarily.