Finish implementation of join both as signed in and external user
Reviewed the before_action methods being called in all controllers. Now
It will always first load the room and then try to load and check if the
user is valid. The user data is indexed in the session according to the
room, so the room has to be loaded before.

The meetings controller was reviewed and improved to accommodate the join
of users without a session using the same join route used by users with
a session. If there's a valid session, it will always be used. Otherwise
it tries to get the user's name from the URL.