Less data in the session, store and fetch it from AppLaunch
Will store only the launch and the expire time in the session. The rest
is taken from the AppLaunch using the launch nonce.

Refactored the before_actions in the controllers. Now the 'omniauth_auth'
in the session is only used on launch, then cleaned up.
The launch action is treated in a special way, that triggers the
authentication. All other routes will just using the session already
created, if any.