Add option to configure an internal bbb server
It's optional. If configured, all API calls made between servers will use
this URL, otherwise they will use the URL in the default endpoint config.

This is useful when running the app in a cluster, so we can use internal
URLs to connect the services and use the external URL just for the calls
that will go to the user's browser (join).