Add BigBlueButton servers, selected for API calls based on consumer keys
New table to register servers that will be selected before each API call
based on the oauth consumer key that is set in the room.

The servers are indexed by this key and have an external endpoint, a
secret and an internal endpoint (optional).

Use rake db:servers:add to add/update servers.