Configure a failure action for omniauth to deal with cookie blockers
The implementation in e80ca28 wasn't fully working in production because
the error happens before SessionsController#create is called. The way to
deal with it is to configure omniauth to call a certain controller/action
on error and deal with the error there.

Had to change the strategy Bbbltibroker a bit to keep the launch nonce
all the way until the last callback made by the oauth, so we can use it
in the retry page.