To retry a failed auth, give the user options to join a conference
The option to launch the room again in another tab wasn't working. The
user needs a full launch, that goes through the broker, for the launch
to work.

Instead, check for current sessions and show the user options to join as
an external user, in case there's a session happening (or close to).