Add attrs and custom launch attrs to disable flags from meetings
Now passing 'custom_allow_wait_moderator' and 'custom_allow_all_moderators'
can enable or disable the edition of the flags 'wait_moderator' and
'all_moderator' in meetings and rooms. If they are not informed, users
are allowed to edit the attributes. If they are informed with 'true', users
are also allowed to edit. If they are informed with any other value,
users will not be allowed to edit the attributes and they will assume
their default values when used.