Search only meetings in the target room when loading them
It was searching for meetings in all rooms for all actions. This wasn't an
issue for actions that validated the room (so user's were not able to
edit a meeting from another room, for example), but was when accessing
open routes, such as /external. You could have always the same room handler
but open any scheduled meeting id and it would work.