Add a repeat attribute to meetings to make them recurring
The attribute can be set so the schedule will be repeated in the future.
When rendering rooms/show all expired recurring meetings are checked and
updated to their next date. The meeting is the same, it just has its start
date updated.
This process can be done in background in the future.

Replaces cf3e8d6.