ELOSP-457 Use oauth_consumer_key when generating room handlers
Use the oauth_consumer_key that is set by the broker in the launch instead
of using parameters that the LMS can change. This makes it more certain
that handlers will be unique for each key used, so different clients
won't mess with others' rooms.

Verified in our production database that this won't change the handlers of
the rooms already there because tool_consumer_instance_guid has already
the same value of oauth_consumer_key.