Send calendar event through OAuth2 authorization on meeting scheduling
When a meeting is succesfuly scheduled, it will use OAuth2 to ask permission
to and send a calendar event to Brightspace.

The oauth strategy is configured in a gem called omniauth-brightspace.

The BrightspaceOauth will store the url, oauth keys (client_id and
client_secret) and is associated with one or more BigbluebuttonServers.