Include a link to the class in the calendar
Moved build_event from BrightspaceController to BrightspaceHelper.

The link is created via API requests and its id is saved on db.
The link is used to create a quicklink, also via API requests, which
is inserted in the Calendar description.

By clicking on the link in the calendar, the user will be redirected
to the external_room_scheduled_meeting_path.