Fix on Abilities when there's no ConsumerConfig
The check for :download_presentation_video was failing if there was no
ConsumerConfig in the database for the target room. Added a few checks to
make sure there will be a default (all users can download the video)
and it will only be changed if there's an specific configuration for it.