Fix error on meeting destruction
The parameter passed in the URL (send_calendar_update_event, etc.) is a
hash_id due to previous changes.
BrightspaceCalendarEvent was storing the scheduled_meeting_id instead of
the scheduled_meeting_hash_id and the event was never found.

It was necessary to change the event model to store the hash_id because the
schedule_meeting object doens't exist in the DB anymore when the
action is called.